+++ Emulator/VICII/VICII.cpp
@@ -50,7 +50,7 @@ VICII::VICII(C64 &ref) : C64Component(ref)
     // Create random background noise pattern
     const size_t noiseSize = 2 * 512 * 512;
     noise = new u32[noiseSize];
-    for (int i = 0; i < noiseSize; i++) {
+    for (size_t i = 0; i < noiseSize; i++) {
         noise[i] = rand() % 2 ? 0xFF000000 : 0xFFFFFFFF;
     }
 }
